+++ src/daft-hash/src/lib.rs
@@ -1,7 +1,14 @@
 #![feature(split_array)]
 
-use std::hash::{BuildHasher, Hasher};
+use std::{
+    hash::{BuildHasher, Hasher},
+    str::FromStr,
+};
 
+use common_error::DaftError;
+#[cfg(feature = "python")]
+use pyo3::prelude::*;
+use serde::{Deserialize, Serialize};
 use sha1::Digest;
 
 pub struct MurBuildHasher {
@@ -42,5 +49,30 @@ impl Hasher for Sha1Hasher {
 
     fn write(&mut self, bytes: &[u8]) {
         self.state.update(bytes);
+    }
+}
+
+/// Format of a file, e.g. Parquet, CSV, JSON.
+#[derive(Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, Copy)]
+#[cfg_attr(feature = "python", pyclass(module = "daft.daft"))]
+pub enum HashFunctionKind {
+    MurmurHash3,
+    XxHash,
+    Sha1,
+}
+
+impl FromStr for HashFunctionKind {
+    type Err = DaftError;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "murmur3" => Ok(Self::MurmurHash3),
+            "xxhash" => Ok(Self::XxHash),
+            "sha1" => Ok(Self::Sha1),
+            _ => Err(DaftError::ValueError(format!(
+                "Hash function {} not found",
+                s
+            ))),
+        }
     }
 }
